+++ src/migrations/m000000_000005_types.php
@@ -1,0 +1,48 @@
+<?php
+/**
+ * m000000_000005_types.php
+ *
+ * PHP version 7.2+
+ *
+ * @author Philippe Gaultier <pgaultier@redcat.io>
+ * @copyright 2010-2019 Redcat
+ * @license https://www.redcat.io/license license
+ * @version XXX
+ * @link https://www.redcat.io
+ * @package blackcube\core\migrations
+ */
+
+namespace blackcube\core\migrations;
+
+use yii\db\Migration;
+
+/**
+ * Class m000000_000005_types
+ *
+ * @author Philippe Gaultier <pgaultier@redcat.io>
+ * @copyright 2010-2019 Redcat
+ * @license https://www.redcat.io/license license
+ * @version XXX
+ * @link https://www.redcat.io
+ * @package blackcube\core\migrations
+ */
+class m000000_000005_types extends Migration
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function safeUp()
+    {
+        $this->alterColumn('{{%types}}', 'route', $this->string()->defaultValue(null));
+        return true;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function safeDown()
+    {
+        $this->alterColumn('{{%types}}', 'route', $this->string()->notNull());
+        return true;
+    }
+}