+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Blackcube Core - CHANGELOG 
 ==========================
 
+Release 1.1.0 June, 10th 2020 <pgaultier@redcat.io>
+---------------------------------------------------
+
+ * Update code to make routing not mandatory in `Type` 
+
 Release 1.0.0 May, 29th 2020 <pgaultier@redcat.io>
 --------------------------------------------------